Refactor map popup
