+++ frontend/src/components/maps/leaflet/LayerPopup/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
 export * from './hooks/useLayerQuery';
+export * from './LayerPopup';
 export * from './LayerPopupContent';
