+++ frontend/src/components/maps/leaflet/LayerPopup/styles.ts
@@ -1,0 +1,17 @@
+import Row from 'react-bootstrap/Row';
+import { Link } from 'react-router-dom';
+import styled from 'styled-components';
+
+export const LayerPopupTitle = styled('div')`
+  padding: 1.6rem;
+  font-weight: 800;
+`;
+
+export const MenuRow = styled(Row)`
+  text-align: center;
+  padding-bottom: 1rem;
+`;
+
+export const StyledLink = styled(Link)`
+  padding: 0 0.4rem;
+`;
